moved ban tests to ban-unbanTest.js
